refactor: moving deps to a plugin dir
