+++ package.json
@@ -15,7 +15,7 @@
     "@web3modal/wagmi": "^3.5.6",
     "dayjs": "^1.11.10",
     "ipfs-http-client": "^60.0.1",
-    "next": "14.0.2",
+    "next": "14.1.0",
     "react": "^18",
     "react-blockies": "^1.4.1",
     "react-dom": "^18",
