+++ pages/proposals/[proposalId]/index.tsx
@@ -1,19 +1,19 @@
 import { usePublicClient, useAccount, useContractWrite } from "wagmi";
 import { useState, useContext, useEffect } from "react";
 import { Address } from "viem";
-import { Proposal } from "@/utils/types";
-import { useProposal } from "@/hooks/useProposal";
-import { useProposalVotes } from "@/hooks/useProposalVotes";
+import { Proposal } from "@/tokenVoting/utils/types";
+import { useProposal } from "@/tokenVoting/hooks/useProposal";
+import { useProposalVotes } from "@/tokenVoting/hooks/useProposalVotes";
 import { Button } from "@aragon/ods";
-import ProposalDescription from "@/components/proposal/description";
-import VotesSection from "@/components/vote/votes-section";
-import ProposalHeader from "@/components/proposal/header";
+import ProposalDescription from "@/tokenVoting/components/proposal/description";
+import VotesSection from "@/tokenVoting/components/vote/votes-section";
+import ProposalHeader from "@/tokenVoting/components/proposal/header";
 import { formatUnits } from "viem";
-import { useUserCanVote } from "@/hooks/useUserCanVote";
-import { TokenVotingAbi } from "@/artifacts/TokenVoting.sol";
-import VoteTally from "@/components/vote/tally";
-import VotingModal from "@/components/vote/voting-modal";
-import ProposalDetails from "@/components/proposal/details";
+import { useUserCanVote } from "@/tokenVoting/hooks/useUserCanVote";
+import { TokenVotingAbi } from "@/tokenVoting/artifacts/TokenVoting.sol";
+import VoteTally from "@/tokenVoting/components/vote/tally";
+import VotingModal from "@/tokenVoting/components/vote/voting-modal";
+import ProposalDetails from "@/tokenVoting/components/proposal/details";
 import { useAlertContext, AlertContextProps } from "@/context/AlertContext";
 import { Else, If, IfCase, Then } from "@/components/if";
 import { PleaseWaitSpinner } from "@/components/please-wait";
