+++ tokenVoting/components/proposal/description.tsx
@@ -1,4 +1,5 @@
-import { Action, Proposal } from "@/utils/types";
+import { Action } from "@/utils/types";
+import { Proposal } from "@/tokenVoting/utils/types";
 import { whatsabi } from "@shazow/whatsabi";
 import { ReactNode, useCallback, useEffect, useState } from "react";
 import { usePublicClient } from "wagmi";
