+++ tokenVoting/components/proposal/details.tsx
