+++ tokenVoting/hooks/useCanCreateProposal.tsx
@@ -1,7 +1,7 @@
 import { Address } from 'viem'
 import { useState, useEffect } from 'react'
 import { useContractReads, useBalance, useAccount } from 'wagmi';
-import { TokenVotingAbi } from '@/artifacts/TokenVoting.sol';
+import { TokenVotingAbi } from '@/tokenVoting/artifacts/TokenVoting.sol';
 
 
 const pluginAddress = ((process.env.NEXT_PUBLIC_PLUGIN_ADDRESS || "") as Address)
